goit-pyweb-hw-13.1/requirements.txt
fastapi
sqlalchemy
psycopg2-binary
pydantic
passlib
python-dotenv
cloudinary
fastapi-limiter
